+++ src/main/java/vo/Profit.java
@@ -1,4 +1,4 @@
 package vo;
 
-public record Profit(int value) {
+public record Profit(double value) {
 }
